Very commented code. Still, I don't get it
- Issue:
  - Create data interactively;
  - add C and WT.
  - Rename C to aC.
  - Names in "Add genotype" are wrong.

It makes no sense. But it eventually works. There is like a lag of one
update. Because if you change the name again, to the same thing (aC) it gets
updated.

But still, a second bug: If you then move to other data, and back, you get "all sub-lists in choices must be named"